[BUGFIX] [Hacker Rank]: Warmup: Mini-Max Sum. [BUGFIX] [Hacker Rank]: Warmup: A Very Big Sum.
Windows "long" type is not enough to store a very big result.

+++ src/lib/exercises/src/hackerrank/warmup/a_very_big_sum.cpp
@@ -9,8 +9,8 @@
 
 namespace hackerrank::warmup {
 
-long aVeryBigSum(const std::vector<long> &ar) {
-  const long INIT_VALUE = 0L;
+long long aVeryBigSum(const std::vector<long> &ar) {
+  const long long INIT_VALUE = 0L;
   return std::accumulate(ar.begin(), ar.end(), INIT_VALUE);
 }
